% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{append_units}
\alias{append_units}
\title{append_units}
\usage{
append_units(x)
}
\arguments{
\item{x}{vector of current names

OUTPUT}
}
\value{
variable names with units
}
\description{
DESCRIPTION
append units to variable names
}
\details{
PARAMETERS
}
