% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_pheno_cov_residual.R
\name{cacti_pheno_cov_residual}
\alias{cacti_pheno_cov_residual}
\title{Residualize peak intensity with covariates}
\usage{
cacti_pheno_cov_residual(file_pheno, file_cov, file_pheno_cov_residual)
}
\arguments{
\item{file_pheno}{Path to phenotype expression matrix.
Rows = features; first column = feature ID; remaining columns = samples.}

\item{file_cov}{Path to covariate matrix.
Rows = covariates; first column = covariate ID; remaining columns = samples.}

\item{file_pheno_cov_residual}{Output filename.
Rows = features; first column = feature ID; remaining columns = samples.}
}
\value{
Invisibly returns the residual matrix (features x samples); writes \code{file_pheno_cov_residual}.
}
\description{
Reads a phenotype matrix (features x samples; first column = feature ID)
and a covariate matrix (covariates x samples; first column = covariate ID)
}
