% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{crt_cacti_request}
\alias{crt_cacti_request}
\title{crt_cacti_request}
\usage{
crt_cacti_request(
  funaction_df,
  country,
  cacti_data = 0,
  volume = 40,
  tag_unfiltered = "_C1",
  tag_filtered = "_C2",
  outfname = "cacti_request.csv"
)
}
\arguments{
\item{funaction_df}{funaction kobo csv data}

\item{country}{full name, e.g., "Switzerland"}

\item{cacti_data}{chemistry results received from cacti (if any)
Default 0}

\item{volume}{approx. volume of the water samples (in ml)
Default 40}

\item{tag_unfiltered}{tag used to identify unfiltered samples
Default "_C1"}

\item{tag_filtered}{tag used to identify filtered samples
Default "_C2"}

\item{outfname}{name of the output file (path/filename)

OUTPUT}
}
\value{
message with filepath and filename
}
\description{
DESCRIPTION
create a csv file to request chemistry analyses
}
\details{
PARAMETERS
}
