% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{fix_varnames}
\alias{fix_varnames}
\title{fix_varnames}
\usage{
fix_varnames(x)
}
\arguments{
\item{x}{vector of current names

OUTPUT}
}
\value{
new simplified names
}
\description{
DESCRIPTION
standardization of variable names
}
\details{
PARAMETERS
}
