% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{get_precision}
\alias{get_precision}
\title{get_precision}
\usage{
get_precision(x)
}
\arguments{
\item{x}{target variable

OUTPUT}
}
\value{
precision of target variable
}
\description{
DESCRIPTION
report the corresponding precision to the variable of
interest
}
\details{
PARAMETERS
}
