% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{get_siteID}
\alias{get_siteID}
\title{get_siteID}
\usage{
get_siteID(df)
}
\arguments{
\item{df}{cleaned cacti dataframe (see read_cacti function)

OUTPUT}
}
\value{
vector with site IDs
}
\description{
DESCRIPTION
extract unique site ID from cleaned cacti dataframe
}
\details{
PARAMETERS
}
