% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{getLOQ}
\alias{getLOQ}
\title{getLOQ}
\usage{
getLOQ(x)
}
\arguments{
\item{x}{target variable

OUTPUT}
}
\value{
precision of target variable
}
\description{
DESCRIPTION
report the corresponding limit of quantification LOQ of the
variable of interest
}
\details{
PARAMETERS
}
