% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{crt_cacti_request}
\alias{crt_cacti_request}
\title{crt_cacti_request}
\usage{
crt_cacti_request(
  funaction_df,
  country,
  cacti_data = 0,
  volume = 40,
  tag_unfiltered = "_C1",
  tag_filtered = "_C2",
  outfname = "cacti_request.csv"
)
}
\arguments{
\item{funaction_df}{\link{character} funaction kobo csv data}

\item{country}{\link{character} full name, e.g., "Switzerland"}

\item{cacti_data}{\link{dataframe} chemistry results received from
cacti (if any)
Default 0}

\item{volume}{\link{numeric} approx. volume of the water samples
(in ml)
Default 40}

\item{tag_unfiltered}{\link{character} tag used to identify
unfiltered samples
Default "_C1"}

\item{tag_filtered}{\link{character} tag used to identify
filtered samples
Default "_C2"}

\item{outfname}{\link{character} name of the output file
(path/filename)
Default "cacti_request.csv"

OUTPUT}
}
\value{
message with filepath and filename
}
\description{
DESCRIPTION
create a csv file to request chemistry analyses
}
\details{
PARAMETERS
}
