% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{read_cacti}
\alias{read_cacti}
\title{read_cacti}
\usage{
read_cacti(filename, path = "", show_units = FALSE)
}
\arguments{
\item{filename}{the original *.xlsx cacti data}

\item{path}{path to filename
DEFAULT "" current directory}

\item{show_units}{show measurement units along variable names
DEFAULT FALSE

OUTPUT}
}
\value{
dataframe, cleaned cacti data
}
\description{
DESCRIPTION
read, prepare and clean cacti data
}
\details{
PARAMETERS
}
