% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{read_cacti}
\alias{read_cacti}
\title{read_cacti}
\usage{
read_cacti(fname, show_units = FALSE, lines_to_skip = 1, sheets = c(2, 3))
}
\arguments{
\item{fname}{the original path/filename *.xlsx cacti data}

\item{show_units}{show measurement units along variable names
DEFAULT FALSE}

\item{lines_to_skip}{skip this many lines when reading the
file
DEFAULT 1}

\item{sheets}{from which sheets to read
DEFAULT c(2,3)

OUTPUT}
}
\value{
dataframe, cleaned cacti data
}
\description{
DESCRIPTION
read, prepare and clean cacti data
}
\details{
PARAMETERS
}
