% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{read_cacti}
\alias{read_cacti}
\title{read_cacti}
\usage{
read_cacti(
  fname,
  show_units = FALSE,
  use_cacti_precision = FALSE,
  lines_to_skip = 1,
  sheets = c(2, 3),
  treat_loq = "I(loq)"
)
}
\arguments{
\item{fname}{\link{character} the original path/filename *.xlsx
cacti data}

\item{show_units}{\link{boolean} show measurement units along
variable names
DEFAULT FALSE}

\item{use_cacti_precision}{\link{boolean} report results using
cacti decimal precision
DEFAULT FALSE}

\item{lines_to_skip}{\link{integer} skip this many lines when
reading the file
DEFAULT 1}

\item{sheets}{\link{numeric vector} from which sheets to read
DEFAULT c(2,3)}

\item{treat_loq}{\link{character} method to deal with values below
the limit of quantification LOQ. Currently, two methods are
supported:
(i)  "loq/2" values less than LOQ are divided by two
(ii) "I(loq)" (DEFAULT) applies identity function to values
below LOQ. If values fall below the limit of detection,
DEFAULT

OUTPUT}
}
\value{
dataframe, cleaned cacti data
}
\description{
DESCRIPTION
read, prepare and clean cacti data
}
\details{
PARAMETERS
}
