% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacti.r
\name{read_cacti}
\alias{read_cacti}
\title{read_cacti}
\usage{
read_cacti(filename, show_units = FALSE)
}
\arguments{
\item{show_units}{show measurement units along variable names
DEFAULT FALSE

OUTPUT}

\item{fname}{the original path/filename *.xlsx cacti data}
}
\value{
dataframe, cleaned cacti data
}
\description{
DESCRIPTION
read, prepare and clean cacti data
}
\details{
PARAMETERS
}
